START OF SHOPPING LIST ADDED, I'M SO TIRED

+++ java/src/main/java/com/techelevator/model/Meal.java
@@ -20,7 +20,7 @@ public class Meal {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "mealid")
     @NotNull
-    private Long mealid;
+    private Integer mealid;
 
     @Column(name = "user_id")
     @NotNull
