+++ java/src/main/java/com/techelevator/model/RecipeMealDTO.java
@@ -5,8 +5,8 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 public class RecipeMealDTO {
 
     @JsonProperty
-    private Long recipeid;
+    private Integer recipeid;
 
     @JsonProperty
-    private Long mealid;
+    private Integer mealid;
 }
